requirements-test.txt
-r requirements.txt
pytest
pytest-django
pytest-cov
